aded api call for floor plans
