+++ components/shared/section-header/index.tsx
@@ -19,7 +19,6 @@ const SectionHeader: FC<SectionHeaderProps> = ({ header, description }) => {
         id: "[data-animated-text]",
         trigger: sectionHeaderContainer.current,
         start: "top 90%",
-        markers: true,
       },
     });
     tl.from(sectionHeaderContainer.current, {
